[FEAT] 모달창 추상화 + 페이지네이션 예외 추가
