+++ src/pages/lobby/elements/roomListDetail/RoomContents.jsx
@@ -21,7 +21,6 @@ const RoomContents = ({ isWaiting, isPrivate }) => {
     navigate(`/game/${roomId}`);
   };
 
-  useEffect(() => {});
   return (
     <StWrapper>
       {status === "loading" && <div>Loading...</div>}
